+++ src/Components/Login/Login.js
@@ -1,9 +1,16 @@
-import React from 'react'
+import React, { useContext } from 'react'
+import { UserStates } from '../../App'
 
 function Login() {
+    const { ModalState } = useContext(UserStates)
+    const { modalOpen, setModalOpen } = ModalState
+
+
+
     return (
         <div>
-            <h1 style={{textAlign:"center"}}>Login here</h1>
+            <h1 style={{ textAlign: "center" }}>Login here</h1>
+            <button onClick={() => setModalOpen(false)}>Modal toggle</button>
         </div>
     )
 }