+++ src/Pages/Pages.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import AdminPanel from '../Components/AdminPanel/AdminPanel'
+
+function Pages() {
+    return (
+        <div>
+            <AdminPanel/>
+        </div>
+    )
+}
+
+export default Pages